downloader.1/requirements.txt
Django~=5.2
yt-dlp~=2025.3.31
celery~=5.5.3
djangorestframework~=3.16.0
django-cors-headers~=4.7.0
redis==5.0.3
